add example code
